Crate alg3_dynamic

An incremental implementation of worst-case optimal joins.

This crate contains functionality to construct timely dataflow computations to compute and maintain the results of complex relational joins under changes to the relations, with worst-case optimality guarantees about the running time.

As an example, consider a stream of directed graph edges (src, dst) where we would like to find all directed cycles of length three. That is, node indentifiers x0, x1, and x2 where the graph contains edges (x0, x1), (x1, x2), and (x2, x0). We can write this query as a relational join on the edge relation edge(x,y), as

cycle_3(x0, x1, x2) := edge(x0, x1), edge(x1, x2), edge(x2, x0)

To determine the set of three-cycles, we could use standard techniques from the database literature to perform the join, typically first picking one attribute (x0, x1, or x2) and performing the join on the two relations containing that attribute, then joining (intersecting) with the remaining relation.

This has the defect that it may perform an amount of work quadratic in the size of edges. Recent work on "worst-case optimal join processing" shows how to get around this problem, by considering multiple relations at the same time.

This crate is a streaming implementation of incremental worst-case optimal join processing. You may indicate a relational query like above, and the crate with synthesize a timely dataflow computation which reports all changes to the occurrences of satisfying assignments to the values. The amount of work performed is no more than the worst-case optimal bound.

Example

fn main () {
    
    // start up a timely dataflow computation
    timely::execute_from_args(std::env::args(), move |root| {

        // build the dataflow graph, return input and output.
        let (edges, probe) = root.scoped(|scope| {

            // construct an input for edge changes.
            let (input, edges) = scope.new_input();

            // index the edge relation by first and second fields.
            let (forward, forward_handle) = edges.concat(&query).index();
            let (reverse, reverse_handle) = edges.concat(&query)
                                                 .map(|((src,dst),wgt)| ((dst,src),wgt))
                                                 .index();

            // construct the motif dataflow subgraph.
            let motif = vec![(0,1), (1,2), (2,0)];
            let cycles = general_motif(motif, &edges, &forward, &reverse);
 
            // count cycles, print, return status of the result.
            let probe = cycles.count()
                              .inspect(|x| println("found: {:?}", x))
                              .probe().0;

            (input, probe)
        });         

        // now supply a stream of changes to the graph.         
        edges.send(((0, 1), 1));
        edges.send(((1, 2), 1));
        edges.send(((2, 0), 1));
        edges.advance_to(1);
        root.step_while(|| probe.lt(&edges.time()));
        // should report `(0, 1, 2)`.

        edges.send(((0, 0), 1));
        edges.advance_to(2);
        root.step_while(|| probe.lt(&edges.time()));
        // should report `(0, 0, 0)`

        edges.send(((0, 1), -1));
        edges.advance_to(2);
        root.step_while(|| probe.lt(&edges.time()));
        // should report `(0, 1, 2)` with a negative update.
    }
}

Modules

motif
Construction of dataflow for general graph motifs.

Structs

Index
A multiversion multimap from Key to Val.

IndexStream
A pair of update stream and update index.

Traits

GenericJoin
Extension method for generic join functionality.

Indexable
Arranges something as an IndexStream.

StreamPrefixExtender
Functionality used by GenericJoin to extend prefixes with new attributes.

Functions

advance
Reports the number of elements satisfing the predicate.