Struct alg3_dynamic::motif::GraphStreamIndex

pub struct GraphStreamIndex<G: Scope, H1: Fn(Node) -> u64, H2: Fn(Node) -> u64> where
    G::Timestamp: Ord + Hash,  {
    pub forward: IndexStream<Node, Node, H1, G::Timestamp>,
    pub reverse: IndexStream<Node, Node, H2, G::Timestamp>,
    // some fields omitted
}

Indices and updates for a graph stream.

Fields
forward: IndexStream<Node, Node, H1, G::Timestamp> reverse: IndexStream<Node, Node, H2, G::Timestamp>
Methods

impl<G: Scope, H1: Fn(Node) -> u64 + 'static, H2: Fn(Node) -> u64 + 'static> GraphStreamIndex<G, H1, H2> where
G::Timestamp: Ord + Hash,

pub fn from(
initially: Stream<G, Edge>,
updates: Stream<G, (Edge, i32)>,
hash1: H1,
hash2: H2
) -> (Self, GraphStreamIndexHandle<G::Timestamp>)

Constructs a new graph stream index from initial edges and an update stream.

pub fn from_separately(
initially_f: Stream<G, Edge>,
initially_r: Stream<G, Edge>,
updates: Stream<G, (Edge, i32)>,
hash1: H1,
hash2: H2
) -> (Self, GraphStreamIndexHandle<G::Timestamp>)

Constructs a new graph stream index from initial edges and an update stream.

pub fn from_separately_static(
initially_f: Stream<G, Edge>,
initially_r: Stream<G, Edge>,
queries: Stream<G, (Edge, i32)>,
hash1: H1,
hash2: H2
) -> (Self, GraphStreamIndexHandle<G::Timestamp>)

Constructs a new graph stream index from initial edges and an update stream.

pub fn track_motif<'a>(
&self,
description: &[(usize, usize)]
) -> Stream<G, (Vec<Node>, i32)> where
G: 'a,

Constructs a dataflow subgraph to track a described motif.

pub fn build_motif<'a>(
&self,
description: &[(usize, usize)]
) -> Stream<G, (Vec<Node>, i32)> where
G: 'a,

Constructs a dataflow subgraph to track a described motif.