Struct alg3_dynamic::Index

pub struct Index<Key: Ord + Hash, Val: Ord, T> { /* fields omitted */ }

A multiversion multimap from Key to Val.

An Index represents a multiversion (Key, Val) relation keyed on the first field. It presently assumes that the keys are dense, and so uses a Vec<State> to maintain per-key state. This could be generalized (and may need to be) to index structures such as e.g. HashMap.

Methods

impl<Key: Ord + Hash + Clone, Val: Ord + Clone, T: Ord + Clone> Index<Key, Val, T>

pub fn new() -> Self

Allocates a new empty index.

pub fn count<P, K, Valid, W>(
&mut self,
data: &mut Vec<(P, u64, u64, W)>,
func: &K,
_valid: &Valid,
ident: u64
) where
K: Fn(&P) -> Key,
Valid: Fn(&T) -> bool,

Updates entries of data to reflect counts in the index.

This method may overwrite entries in data to replace the second and third fields with the count of extensions this index would propose and ident, respectively. This overwrite happens if the counts proposed here would be smaller than what is currently recorded in the tuple.

pub fn propose<P, K, Valid, W>(
&mut self,
data: &mut Vec<(P, Vec<Val>, W)>,
func: &K,
valid: &Valid
) where
K: Fn(&P) -> Key,
Valid: Fn(&T) -> bool,

Proposes extensions for prefixes based on the index.

pub fn intersect<P, F, Valid, W>(
&mut self,
data: &mut Vec<(P, Vec<Val>, W)>,
func: &F,
valid: &Valid
) where
F: Fn(&P) -> Key,
Valid: Fn(&T) -> bool,

Restricts extensions for prefixes to those found in the index.

pub fn merge_to(&mut self, time: &T)

Commits updates up to and including time.

This merges any differences with time less or equal to time, and should probably only be called once the user is certain to never require such a distinction again. These differences are not yet compacted, they've just had their times stripped off.

This operation is important to ensure that self.diffs doesn't grow too large, as our strategy for keeping it sorted is to re-sort it whenever we add data. If it grew without bound this would be pretty horrible. In principle, this operation also allows us to consolidate the representation, if we have updates which update the same value (potentially cancelling).

pub fn update(&mut self, time: T, updates: &mut Vec<((Key, Val), i32)>)

Introduces a collection of updates at various times.

These updates will now be reflected in all queries against the index, at or after the indicated logical time.

pub fn initialize(&mut self, initial: &mut Vec<Vec<(Key, Val)>>)

Sets an initial collection of positive counts, which we can compact.