Trait alg3_dynamic::StreamPrefixExtender

pub trait StreamPrefixExtender<G: Scope> {
    type Prefix: Data;
    type Extension: Data;
    fn count(
        &self, 
        _: Stream<G, (Self::Prefix, u64, u64, i32)>, 
        _: u64
    ) -> Stream<G, (Self::Prefix, u64, u64, i32)>;

    fn propose(
        &self, 
        _: Stream<G, (Self::Prefix, i32)>
    ) -> Stream<G, (Self::Prefix, Vec<Self::Extension>, i32)>;

    fn intersect(
        &self, 
        _: Stream<G, (Self::Prefix, Vec<Self::Extension>, i32)>
    ) -> Stream<G, (Self::Prefix, Vec<Self::Extension>, i32)>;
}

Functionality used by GenericJoin to extend prefixes with new attributes.

These methods are used in GenericJoin's extend method, and may not be broadly useful elsewhere.

Associated Types

type Prefix: Data

The type of data to extend.

type Extension: Data

The type of the extentions.

Required Methods

fn count(
&self,
_: Stream<G, (Self::Prefix, u64, u64, i32)>,
_: u64
) -> Stream<G, (Self::Prefix, u64, u64, i32)>

Updates each prefix with an upper bound on the number of extensions for this relation.

fn propose(
&self,
_: Stream<G, (Self::Prefix, i32)>
) -> Stream<G, (Self::Prefix, Vec<Self::Extension>, i32)>

Proposes each extension from this relation.

fn intersect(
&self,
_: Stream<G, (Self::Prefix, Vec<Self::Extension>, i32)>
) -> Stream<G, (Self::Prefix, Vec<Self::Extension>, i32)>

Restricts proposals by those this relation would propose.