Struct alg3_dynamic::motif::GraphStreamIndex

pub struct GraphStreamIndex<G: Scope> where
    G::Timestamp: Ord + Hash,  {
    pub forward: IndexStream<G>,
    pub reverse: IndexStream<G>,
    // some fields omitted
}

Indices and updates for a graph stream.

Fields
forward: IndexStream<G> reverse: IndexStream<G>
Methods

impl<G: Scope> GraphStreamIndex<G> where
G::Timestamp: Ord + Hash,

fn from(
initially: Stream<G, (u32, u32)>,
updates: Stream<G, ((u32, u32), i32)>
) -> (Self, GraphStreamIndexHandle<G::Timestamp>)

Constructs a new graph stream index from initial edges and an update stream.

fn from_separately(
initially_f: Stream<G, (u32, u32)>,
initially_r: Stream<G, (u32, u32)>,
updates: Stream<G, ((u32, u32), i32)>
) -> (Self, GraphStreamIndexHandle<G::Timestamp>)

Constructs a new graph stream index from initial edges and an update stream.

fn track_motif<'a>(
&self,
description: &[(usize, usize)]
) -> Stream<G, (Vec<u32>, i32)> where
G: 'a,

Constructs a dataflow subgraph to track a described motif.