Struct alg3_dynamic::Index

pub struct Index<Key: Ord + Hash, T> { /* fields omitted */ }

A multiversion multimap from Key to Val.

An Index represents a multiversion (Key, Val) relation keyed on the first field. It presently assumes that the keys are dense, and so uses a Vec<State> to maintain per-key state. This could be generalized (and may need to be) to index structures such as e.g. HashMap.

Methods

impl<Key: Ord + Hash + Copy, T: Ord + Clone> Index<Key, T>

fn new() -> Index<Key, T>

Allocates a new empty index.

fn count<P, K, V>(
&mut self,
data: &mut Vec<(P, u64, u64, i32)>,
func: &K,
_valid: &V,
ident: u64
) where
K: Fn(&P) -> Key,
V: Fn(&T) -> bool,

Updates entries of data to reflect counts in the index.

fn propose<P, K, V>(
&mut self,
data: &mut Vec<(P, Vec<u32>, i32)>,
func: &K,
valid: &V
) where
K: Fn(&P) -> Key,
V: Fn(&T) -> bool,

Proposes extensions for prefixes based on the index.

fn intersect<P, F, V>(
&mut self,
data: &mut Vec<(P, Vec<u32>, i32)>,
func: &F,
valid: &V
) where
F: Fn(&P) -> Key,
V: Fn(&T) -> bool,

Restricts extensions for prefixes to those found in the index.

fn merge_to(&mut self, time: &T)

Commits updates up to and including time.

This merges any differences with time less or equal to time, and should probably only be called once the user is certain to never require such a distinction again. These differences are not yet compacted, they've just had their times stripped off.

This operation is important to ensure that self.diffs doesn't grow too large, as our strategy for keeping it sorted is to re-sort it whenever we add data. If it grew without bound this would be pretty horrible. In principe, this operation also allows us to consolidate the representation, if we have updates which update the same value (potentially cancelling).

fn update(&mut self, time: T, updates: &mut Vec<(Key, (u32, i32))>)

Introduces a collection of updates at various times.

These updates will now be reflected in all queries against the index, at or after the indicated logical time.

fn initialize(&mut self, initial: &mut Vec<Vec<(Key, u32)>>)

Sets an initial collection of positive counts, which we can compact.