Struct alg3_dynamic::IndexStream

pub struct IndexStream<K: Ord + Hash + Clone, V: Ord + Clone, H: Fn(K) -> u64, T: Timestamp> {
    pub handle: ProbeHandle<T>,
    pub index: Rc<RefCell<Index<K, V, T>>>,
    pub hash: Rc<H>,
}

An index materialized from streamed updates.

The data are updates to (key, value) pairs, indicated by an associated signed integer. The index stream provides information about update times it has completely accepted via its handle field, which is a timely dataflow probe handle, and can be interrogated about whether outstanding times might still exist less than any query time. There is also a function hash from the key type K to u64 values to indicate how the data are partitioned, so that users can align their query streams.

Fields
handle: ProbeHandle<T>

Times completely absorded into the index.

index: Rc<RefCell<Index<K, V, T>>>

The index itself.

hash: Rc<H>

A map from keys to u64 values, for distribution.

Methods

impl<K: Ord + Hash + Clone, V: Ord + Clone, H: Fn(K) -> u64, T: Timestamp + Ord> IndexStream<K, V, H, T>

pub fn extend_using<P, L, F>(
&self,
logic: L,
func: F
) -> Rc<IndexExtender<K, V, T, P, L, H, F>> where
L: Fn(&P) -> K + 'static,
F: Fn(&T, &T) -> bool + 'static,

Extends an IndexStream using the supplied functions.

The logic function maps prefixes to index keys. The func function compares timestamps, acting as either lt or le depending on the need.

pub fn from<G>(
hash: H,
initially: &Stream<G, (K, V)>,
updates: &Stream<G, ((K, V), i32)>
) -> Self where
G: Scope<Timestamp = T>,
K: ExchangeData,
V: ExchangeData,
T: Hash,
H: 'static,

Constructs an IndexStream from initial data and update stream.

Neither the initial stream nor the update stream are required to produce data. The index can be static with no changes, or wholy dynamic with no starting data, or a mix of both. If neither stream has any data, you are probably using the wrong abstraction (though it will still work correctly).