Struct alg3_dynamic::IndexStream

pub struct IndexStream<G: Scope> where
    G::Timestamp: Ord,  {
    pub stream: Stream<G, (u32, (u32, i32))>,
    // some fields omitted
}

A pair of update stream and update index.

The stream itself is empty, but its frontier can be used as a guarantee that the index contents are locked down for certain times. The index is only available through the get_index() method, to ensure that you don't write to it here.

Fields
stream: Stream<G, (u32, (u32, i32))>
Methods

impl<G: Scope> IndexStream<G> where
G::Timestamp: Ord,

fn extend_using<P, L, F>(
&self,
logic: L,
func: F
) -> Rc<IndexExtender<G, P, L, F>> where
L: Fn(&P) -> u64 + 'static,
F: Fn(&G::Timestamp, &G::Timestamp) -> bool + 'static,

Extends an IndexStream using the supplied functions.

The logic function maps prefixes to index keys. The func function compares timestamps, acting as either lt or le depending on the need.