Trait alg3_dynamic::StreamPrefixExtender

pub trait StreamPrefixExtender<G: Scope, W: Data> {
    type Prefix: Data;
    type Extension: Data;
    fn count(
        &self, 
        _: Stream<G, (Self::Prefix, u64, u64, W)>, 
        _: u64
    ) -> Stream<G, (Self::Prefix, u64, u64, W)>;

    fn propose(
        &self, 
        _: Stream<G, (Self::Prefix, W)>
    ) -> Stream<G, (Self::Prefix, Vec<Self::Extension>, W)>;

    fn intersect(
        &self, 
        _: Stream<G, (Self::Prefix, Vec<Self::Extension>, W)>
    ) -> Stream<G, (Self::Prefix, Vec<Self::Extension>, W)>;
}

Functionality used by GenericJoin to extend prefixes with new attributes.

These methods are used in GenericJoin's extend method, and may not be broadly useful elsewhere.

Associated Types

type Prefix: Data

The type of data to extend.

type Extension: Data

The type of the extentions.

Loading content...
Required methods

fn count(
&self,
_: Stream<G, (Self::Prefix, u64, u64, W)>,
_: u64
) -> Stream<G, (Self::Prefix, u64, u64, W)>

Updates each prefix with an upper bound on the number of extensions for this relation.

fn propose(
&self,
_: Stream<G, (Self::Prefix, W)>
) -> Stream<G, (Self::Prefix, Vec<Self::Extension>, W)>

Proposes each extension from this relation.

fn intersect(
&self,
_: Stream<G, (Self::Prefix, Vec<Self::Extension>, W)>
) -> Stream<G, (Self::Prefix, Vec<Self::Extension>, W)>

Restricts proposals by those this relation would propose.

Loading content...